Fix tooltip Z-layers, add layers_swallow
To combat future endeavors with trying to make the tooltips appear where
they should, I've created an entire new Tooltips layer in very room and
created a persistent obj_tooltips object which uses a global queue to
draw every needed tooltip at the same Z-layer.

Since this newly created obj_tooltips object is the second one to be in
need of being reassigned to the same-named layer every time a room
changes, I've put together a small function (layers_swallow) to reassign
the layers for obj_tooltips and boss (which used to be the only object
in need of such reassignments).