Add Surface Maps as tradeoff of memory for CPU
Bobileusz's bullets require very frequent creation and freeing of
surfaces of various sizes, which is a big strain on the CPU.
I could see even more kinds of objects requiring this type of
allocation.

The solution is very simple, a ds_list of square surfaces that can be
used for anything. A parallel ds_list keeps track of which surfaces are
available and which are being used, and you can reuse the same surfaces
between any number of objects without freeing them and allocating anew.

There is a memory tradeoff, because the available surfaces don't have
optimized sizes, so there's quite a lot of them that are too big, but
as long as this feature is used reasonably, it can significantly
increase performance.