Use rectangular masks for Bobileusz projectiles
This is a huge optimization step, lifting average FPS from around 60 up
to 700. The previous collision detection method involved creating and
deleting sprite resources with precise masks, which was completely
inefficient. The current method uses a rectangular mask and stretches it
to match a rectangle that's approximately covering the light rays.
It's surprisingly accurate and at a very low CPU cost.