Insert particle class suffix to all macros
When spawning particles or doing other particle-related things, it's too
easy to use the wrong particle class (ult, lt, pro) with a particle
system/type of a different class, which leads to very strange results
and takes a while to realize what's wrong. Until now.

Now it's impossible to make a mistake, because all particle system and
type macros end with the abbreviation of their particle class. This is a
small change, but it's going to make working with particles a lot
easier.