Upgrade surface maps to surface pages
The issue with surface maps is that sometimes we need a lot of surfaces
of different types, and surface maps give out surfaces only based on
whether the size is big enough, so you could have a bullet needing a 3x3
surface, but occupying a 3000x3000 surface (theoretically).

The simple solution is surface pages, so basically we're splitting
surface maps into separate lists. Each page has the same functionality
as surface pages, but now objects that need surfaces of similar sizes
should be assigned the same surface page, and ones that need different
sizes can have their own surface pages. That way the memory footprint is
minimized.