Stabilize save/load with all new Bobileusz stuff
I DEFINITELY need to revamp the saving and loading system, there's a lot
of improvements to be had, to name a few:

1) Game version should be unencrypted in the save file, because there
should be an easy way to display it in the slot selection menu
2) Save file should be order agnostic, e.g. each section delimited by
"###" should be preceded with some sort of save chunk ID that signifies
what kind of data is stored there. The current system is chronological,
meaning inserting things in the middle or anywhere in general is a huge
pain. Maybe each save chunk ID could also hold a game version or
something, that could make backwards compatibility way easier to
maintain.