Revamp entire save system, major reconstructions
Fortunately, I've scattered a handful of useful comments about most of
the changes, because it's been way too many for me to remember and list
them in a complete shape and form here:

- saving and loading slots is now an order of magnitude faster and more
  stable (versions 1.2.0 beta and later)
- changed sound_priority enum to a ds_map for convenience
- added version comparison scripts to easily find out which game version
  precedes which, etc.

Old slot loading code for versions 1.1.0 and 1.1.1 is still available,
but it's been separated into obj_slot_loading_legacy. THIS HASN'T BEEN
TESTED yet, I will run more extensive testing on older slots in the
future.