Convert project to GMS 2.3
GameMaker: Studio has released a new major version, 2.3, which alters
how some things are saved in the filesystem. This snapshot ONLY CONVERTS
the old project to the latest format, without any changes in the actual
game code whatsoever. It's possible that this change will introduce new
bugs which will have to be fixed later.