Fix slot version checking when loading things
I made a mistake of comparing game version instead of slot version in
one spot. To make such errors easier to spot in the future, I've added
a line to each of the loading scripts that unpacks the slot version
argument into a temporary variable (even though it's not used in most of
them).