Fix loading glacier and spawn ID crashes
The problem was, Bobileusz glacier's spawn value was set to obj_player
instead of an appropriate INSTANCE id. Implicit conversion is now made
when loading slots. Projectiles, minions, frags and charges now search
for the current obj_player instance if the saved ID is inappropriate.
This is only to prevent crashes, the preferred way is saving IDs
properly.