Fix having to press twice to start a run
This issue was caused by the scr_PlayerUpdateGroups script, which would
unload all audio groups and then load only the needed one for the
current player. Little did I know when initially writing it, unloading
an audio group is ALSO and asynchronous action which takes additional
time, but for some reason GMS does not capture it in the async_load
variable (so basically there's no way of telling when an audio group has
been unloaded, exactly). For this reason I've changed the functionality
a bit:

Instead of the former method, only unload the necessary audio groups,
then IF the target character's audio group is unloaded, load it as
always, but if it's not, start the run immediately without any async
event happening. To make this tidy I've simply extracted the code that
starts the run from the async event into a script (scr_RunStart), and
I'm calling it both from the async event and from scr_PlayerUpdateGroups
in that second mentioned case.

No rigorous testing was performed, but I don't see how anything could
break from this change.