Fix home script error, purge obsolete functions
The home script/function used to overwrite argument[0]. It is not
mentioned in the manual that argument[n] is read-only, but it causes a
runtime error pretty frequently, so it needed to be changed. Shame on
you, YoyoGames.

Replaced array_length_1d/array_length_2d/array_height_2d functions with
array_length, as specified in the newer standard.