Tweak Gingerbread Man blue bolts' lifespan
The blue bolts now spawn with a much smaller, random lifespan. This is
done to limit the number of blue bolts on the screen, as sometimes, when
there's a lot of gingerbread men, it gets pretty crazy and they all used
to have a 30 second (virtually indepletable) lifespan, much like the
yellow bolts.