Buffdate Evilflame
- lower charging & cooldown
- Dual Clone can now be "cancelled" by spending a power point, which
causes 6 rapid charges to be shot, and the sum of HP is preserved. This
is very nice, as it gets rid of the issue of having power points, but
being unable to spend them (when in dual mode), and it also provides
incentive for strategic planning when to cancel the clone, because it
gives a great attack benefit.
- activating dual clone now grants +50 HP (+25 for each ship). This
change was made to counter the negative effects of Dual Clone (it's very
easy to lose HP). This added bonus, and the new "cancel" attack are
meant to give more fun to the user, and buff Evilflame so that he can
enter high-point rounds (+100k) with relative ease, like every other
character.