Fix issues with SFX sounds
Apparently GMS2 now does not let you manually order the ids of sfx,
sprites, etc. Luckily for me, the ids of sprites are all correct, but
the sfx aren't, which made it necessary to add a new system for
determining which sfx belongs to which character. Hopefully the same
won't have to be done for sprites later.